wip - add roles & cleanup resource structure
