+++ rust/operator-binary/src/main.rs
@@ -49,11 +49,9 @@ mod config;
 mod controller_commons;
 mod crd;
 mod druid_connection_controller;
-mod listener;
 mod operations;
 mod product_logging;
-mod rbac;
-mod service;
+mod resources;
 mod superset_controller;
 mod util;
 mod webhooks;
