+++ rust/operator-binary/src/resources/mod.rs
@@ -1,0 +1,6 @@
+pub mod configmap;
+pub mod deployment;
+pub mod listener;
+pub mod rbac;
+pub mod service;
+pub mod statefulset;
