+++ rust/operator-binary/src/resources/service.rs
